Encapsulate progress measurement in a shared function
